src/requirements.txt
pathspec==0.12.1
openai==0.27.0
python-dotenv==1.0.1